refactor(deprecate): extract callables.py engine
- move the callable engine (`packing`, call-plan, warning helpers) into new `src/deprecate/callables.py`; `deprecated_callable` is now the canonical implementation and rejects class sources with `TypeError`
- rewrite `deprecated` as a thin dispatcher that delegates callables to `deprecated_callable` and classes to `deprecated_class`, reversing the earlier direction so there is no cycle when `deprecated` acts as the all-in-one front door
- reduce `deprecation.py` to a backward-compat shim re-exporting `deprecate.callables` (public and internal names listed in `__all__`)
- parameterize `_warn_stacking_misconfiguration` with `stacklevel` (caller passes `_stacklevel + 1`) so misconfig warnings still point at user code through the extra dispatcher frame
- retarget the `test_property` mock patch, api-reference autodoc paths, and CONTRIBUTING/AGENTS module references to `callables.py`

---
